+++ test/acceptance/familio_marriage_test.go
@@ -5,19 +5,14 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/knownvalue"
+	"github.com/hashicorp/terraform-plugin-testing/plancheck"
 	"github.com/hashicorp/terraform-plugin-testing/statecheck"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/hashicorp/terraform-plugin-testing/tfjsonpath"
 )
 
 func TestAccMarriage_basic(t *testing.T) {
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testAccPreCheck(t) },
-		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
-		CheckDestroy:             checkPersonsDestroyed(t),
-		Steps: []resource.TestStep{
-			{
-				Config: `
+	const couple = `
 resource "familio_person" "husband" {
   first_name = "АкцТест"
   last_name  = "Мужев"
@@ -31,7 +26,14 @@ resource "familio_person" "wife" {
   gender     = "female"
   privacy    = "invisible"
 }
-
+`
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testProtoV6ProviderFactories,
+		CheckDestroy:             checkPersonsDestroyed(t),
+		Steps: []resource.TestStep{
+			{
+				Config: couple + `
 resource "familio_marriage" "test" {
   partners      = [familio_person.husband.uuid, familio_person.wife.uuid]
   marriage_date = { year = 1875, month = 5, day = 12 }
@@ -40,6 +42,25 @@ resource "familio_marriage" "test" {
 					statecheck.ExpectKnownValue("familio_marriage.test", tfjsonpath.New("partners"), knownvalue.SetSizeExact(2)),
 					statecheck.ExpectKnownValue("familio_marriage.test", tfjsonpath.New("marriage_date").AtMapKey("year"), knownvalue.Int64Exact(1875)),
 					statecheck.ExpectKnownValue("familio_marriage.test", tfjsonpath.New("uuid"), knownvalue.NotNull()),
+				},
+			},
+			{
+				// Editing the date and adding a comment must be an in-place update
+				// (the underlying wedding event is rebuilt), NOT a replacement.
+				Config: couple + `
+resource "familio_marriage" "test" {
+  partners      = [familio_person.husband.uuid, familio_person.wife.uuid]
+  marriage_date = { year = 1876 }
+  comment       = "повторное оглашение"
+}`,
+				ConfigPlanChecks: resource.ConfigPlanChecks{
+					PreApply: []plancheck.PlanCheck{
+						plancheck.ExpectResourceAction("familio_marriage.test", plancheck.ResourceActionUpdate),
+					},
+				},
+				ConfigStateChecks: []statecheck.StateCheck{
+					statecheck.ExpectKnownValue("familio_marriage.test", tfjsonpath.New("marriage_date").AtMapKey("year"), knownvalue.Int64Exact(1876)),
+					statecheck.ExpectKnownValue("familio_marriage.test", tfjsonpath.New("comment"), knownvalue.StringExact("повторное оглашение")),
 				},
 			},
 			{